requirements.txt
beautifulsoup4==4.12.2
mysql-connector-python==8.0.33
requests==2.29.0
